Move bootc tests to new ImageMode component
